requirements.txt
attrs>=19.3.0
requests>=2.23.0
